Refactoting: usePostModal customHook
