+++ src/pages/SignUp/index.tsx
@@ -1,17 +1,15 @@
 import { Navigate, Route, Routes } from 'react-router-dom';
 import NavSection from '@/components/signUp/NavSection';
-import { useState } from 'react';
-import { Address } from 'react-daum-postcode';
 import PostModal from '@/components/signUp/PostModal';
 import { SignUpMemberData, SignUpOwnerData } from '@/types/userData';
 import OwnerSection from '@/components/signUp/OwnerSection';
 import MemberSection from '@/components/signUp/MemberSection';
 import axios from 'axios';
 import { handleToastError } from '@/utils/handleToast';
+import usePostModal from '@/hooks/usePostModal';
 
 const SignUp = () => {
-  const [roadAddress, setRoadAddress] = useState('');
-  const [showPostModal, setShowPostModal] = useState(false);
+  const { showPostModal, openPostModal, closePostModal, handleRoadAddress, roadAddress } = usePostModal();
 
   const sendSignUpData = async ({ data, isOwner }: { data: SignUpMemberData | SignUpOwnerData; isOwner: boolean }) => {
     try {
@@ -21,20 +19,6 @@ const SignUp = () => {
     } catch (error) {
       handleToastError(error);
     }
-  };
-
-  const handleRoadAddress = (address: Address) => {
-    console.log(address);
-    setRoadAddress(() => address.roadAddress);
-    closePostModal();
-  };
-
-  const openPostModal = () => {
-    setShowPostModal(() => true);
-  };
-
-  const closePostModal = () => {
-    setShowPostModal(() => false);
   };
 
   return (
